Move require ports reducer
close DCOS-14729
