+++ plugins/services/src/js/reducers/serviceForm/JSONReducers/__tests__/RequirePorts-test.js
@@ -1,8 +1,8 @@
 const RequirePorts = require('../RequirePorts');
-const Batch = require('../../../../../../../src/js/structs/Batch');
-const Transaction = require('../../../../../../../src/js/structs/Transaction');
+const Batch = require('../../../../../../../../src/js/structs/Batch');
+const Transaction = require('../../../../../../../../src/js/structs/Transaction');
 const {SET} =
-  require('../../../../../../../src/js/constants/TransactionTypes');
+  require('../../../../../../../../src/js/constants/TransactionTypes');
 
 describe('RequirePorts', function () {
   describe('#JSONReducer', function () {
